+++ src/Database.java
@@ -82,6 +82,20 @@ public class Database {
 										//if the data to import pertains to a doctor
 										if(importRole.compareTo("doctor") == 0)
 										{
+											//read in doctor availability
+											InputStream st = getClass().getClassLoader().getResourceAsStream("dbase/"+importUsername+"Avail.txt");
+											sc = new Scanner(st);
+											
+											ArrayList<String> tempp = new ArrayList<>(5);
+											while(sc.hasNextLine()) 
+											{
+												
+												String time = sc.nextLine();			//read in entire line
+												
+												tempp.add(time);
+											}
+											
+											
 											//create a temporary doctor object using database information
 											DoctorModel temp = new DoctorModel(importUsername, importPassword, importName
 													, importDepartment, importAssignedNurseUsernames, importAssigPat);
@@ -100,23 +114,41 @@ public class Database {
 										
 										else if(importRole.compareTo("patient")== 0)
 										{
-											//get Patient record
-											try {
-												//https://stackoverflow.com/questions/3891375/how-to-read-a-text-file-resource-into-java-unit-test?noredirect=1&lq=1
-												String record = new String(getClass().getClassLoader()
-														.getResourceAsStream("dbase/"+importUsername+".txt").readAllBytes());
+												//get Patient appt times
+												InputStream st = getClass().getClassLoader().getResourceAsStream("dbase/"+importUsername+"Appt.txt");
+													sc = new Scanner(st);
+													
+													ArrayList<String> tempp = new ArrayList<>(5);
+													while(sc.hasNextLine()) 
+													{
+														
+														String time = sc.nextLine();			//read in entire line
+														String[] split2 = time.split("\t");		//regex split into arrow on tabs
+														
+														String docName = split2[0];
+														String apptTime = split2[1];
+														
+														tempp.add(docName);
+														tempp.add(apptTime);
+													}
 												
-												//create a patient internally
-												PatientModel temp = new PatientModel(importUsername, importPassword, importName
-														, importAddress, importPhoneNumber, importEmail, importAmountDue
-														, importDob, importBlood, importSex, record);
+												//get Patient record notes
+												try {
+													//https://stackoverflow.com/questions/3891375/how-to-read-a-text-file-resource-into-java-unit-test?noredirect=1&lq=1
+													String record = new String(getClass().getClassLoader()
+															.getResourceAsStream("dbase/"+importUsername+".txt").readAllBytes());
+													
+													//create a patient internally
+													PatientModel temp = new PatientModel(importUsername, importPassword, importName
+															, importAddress, importPhoneNumber, importEmail, importAmountDue
+															, importDob, importBlood, importSex, record, tempp);
+													
+													users.put(importUsername, temp);
+												} catch (IOException e) {
+													System.out.println("Could not open file");
+													e.printStackTrace();
+												}
 												
-												users.put(importUsername, temp);
-											} catch (IOException e) {
-												System.out.println("Could not open file");
-												e.printStackTrace();
-											}
-											
 											
 										}